+++ LibraryView.m
@@ -208,7 +208,7 @@
   {
     NSArray* previousSelectedItems = [self itemsAtRowIndexes:[self selectedRowIndexes]];
     NSPoint point = [self convertPoint:[theEvent locationInWindow] fromView:nil];
-    int row = [self rowAtPoint:point];
+    NSInteger row = [self rowAtPoint:point];
     id candidateToSelection = [self itemAtRow:row];
     if (![previousSelectedItems containsObject:candidateToSelection])
       [self selectRowIndexes:[NSIndexSet indexSetWithIndex:row] byExtendingSelection:NO];
@@ -217,8 +217,8 @@
     {
       [super mouseDown:theEvent];
       NSPoint pointInView = [self convertPoint:[theEvent locationInWindow] fromView:nil];
-      int row = [self rowAtPoint:pointInView];
-      int column = [self columnAtPoint:pointInView];
+      NSInteger row = [self rowAtPoint:pointInView];
+      NSInteger column = [self columnAtPoint:pointInView];
       NSRect rect = ((row >= 0) && (column >= 0)) ? [self frameOfCellAtColumn:column row:row] : NSZeroRect;
       NSRect imageFrame = NSZeroRect;
       NSRect titleFrame = NSZeroRect;
@@ -280,7 +280,7 @@
   else if ([[self window] isKeyWindow])//if (NSPointInRect(location, [clipView bounds]))
   {
     location = [self convertPoint:location fromView:clipView];
-    int row = [self rowAtPoint:location];
+    NSInteger row = [self rowAtPoint:location];
     id libraryItem = (row >= 0) && (row < [self numberOfRows]) ? [self itemAtRow:row] : nil;
     NSImage* image = nil;
     NSColor* backgroundColor = nil;
@@ -298,7 +298,7 @@
 
 -(void) cancelOperation:(id)sender
 {
-  int editedRow = [self editedRow];
+  NSInteger editedRow = [self editedRow];
   if (editedRow >= 0)
   {
     LibraryItem* libraryItem = [self itemAtRow:editedRow];
@@ -328,7 +328,7 @@
 
 -(void) edit:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if (selectedRow >= 0)
     [self editColumn:0 row:selectedRow withEvent:nil select:YES];
 }
@@ -369,7 +369,7 @@
   }
   else //if we are going down after an upwards selection, deselect last selected item
   {
-    unsigned int firstIndex = [selectedRowIndexes firstIndex];
+    NSUInteger firstIndex = [selectedRowIndexes firstIndex];
     [self deselectRow:firstIndex];
   }
 }
@@ -378,7 +378,7 @@
 -(void) moveUpAndModifySelection:(id)sender
 {
   //selection to up
-  unsigned int lastSelectedRow   = [self selectedRow];
+  NSUInteger lastSelectedRow   = [self selectedRow];
   NSIndexSet* selectedRowIndexes = [self selectedRowIndexes];
   if (lastSelectedRow == [selectedRowIndexes firstIndex]) //if the selection is going up, and up, increase it
   {
@@ -388,7 +388,7 @@
   }
   else //if we are going up after an downwards selection, deselect last selected item
   {
-    unsigned int lastIndex = [selectedRowIndexes lastIndex];
+    NSUInteger lastIndex = [selectedRowIndexes lastIndex];
     [self deselectRow:lastIndex];
   }
 }
@@ -396,7 +396,7 @@
 
 -(void) moveUp:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if (selectedRow > 0)
     --selectedRow;
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
@@ -406,7 +406,7 @@
 
 -(void) moveDown:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if ((selectedRow >= 0) && (selectedRow+1 < [self numberOfRows]))
     ++selectedRow;
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
@@ -434,7 +434,7 @@
   nextSelectedItem = nextSelectedItem ? nextSelectedItem :
     [[selectedItems lastObject] prevBrotherWithParentSelector:@selector(parent) childrenSelector:@selector(childrenOrdered) rootNodes:[self->libraryController rootItems]];
   nextSelectedItem = nextSelectedItem ? nextSelectedItem : [[selectedItems lastObject] parent];
-  unsigned int nbSelectedItems = [selectedItems count];
+  NSUInteger nbSelectedItems = [selectedItems count];
   NSMutableSet* parentOfSelectedItems = [NSMutableSet setWithCapacity:[selectedItems count]];
   NSEnumerator* enumerator = [selectedItems objectEnumerator];
   LibraryItem* libraryItem = nil;
@@ -591,13 +591,13 @@
   LibraryItem* selectedItem = [self selectedItem];
   LibraryGroupItem* parentOfSelectedItem = (LibraryGroupItem*)[selectedItem parent];
   NSArray* brothers = !parentOfSelectedItem ? [self->libraryController rootItems] : [parentOfSelectedItem childrenOrdered];
-  int childIndex = !selectedItem ? [[self dataSource] outlineView:self numberOfChildrenOfItem:nil] :
-                   ((int)[brothers indexOfObject:selectedItem]+1);
+  NSInteger childIndex = !selectedItem ? [[self dataSource] outlineView:self numberOfChildrenOfItem:nil] :
+                   ((NSInteger)[brothers indexOfObject:selectedItem]+1);
   [self pasteContentOfPasteboard:pasteboard onItem:parentOfSelectedItem childIndex:childIndex];
 }
 //end paste:
 
--(BOOL) pasteContentOfPasteboard:(NSPasteboard*)pasteboard onItem:(id)item childIndex:(int)index
+-(BOOL) pasteContentOfPasteboard:(NSPasteboard*)pasteboard onItem:(id)item childIndex:(NSInteger)index
 {
   BOOL result = NO;
   
@@ -682,7 +682,7 @@
       }
     }//end for each "new" libraryItem
     [brothers insertObjectsFromArray:libraryItems atIndex:(index == NSOutlineViewDropOnItemIndex) ?
-      [brothers count] : (unsigned)index];
+      [brothers count] : index];
     NSUInteger nbBrothers = [brothers count];
     while(nbBrothers--)
       [[brothers objectAtIndex:nbBrothers] setSortIndex:nbBrothers];
