+++ NSMutableArrayExtended.m
@@ -18,7 +18,7 @@
 //end safeAddObject:
 
 //inserts another array's content at a given index
--(void) insertObjectsFromArray:(NSArray *)array atIndex:(int)index
+-(void) insertObjectsFromArray:(NSArray *)array atIndex:(NSInteger)index
 {
   NSEnumerator* enumerator = [array objectEnumerator];
   NSObject* entry = nil;
@@ -27,7 +27,7 @@
 }
 //end insertObjectsFromArray:atIndex:
 
--(void) moveObjectsAtIndices:(NSIndexSet*)indices toIndex:(unsigned int)index
+-(void) moveObjectsAtIndices:(NSIndexSet*)indices toIndex:(NSUInteger)index
 {
   NSArray* objectsToMove = [self objectsAtIndexes:indices];
   NSUInteger shift = 0;
